# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SessãoInk API",
    "description": "API de gestão para tatuadores e estúdios de tatuagem",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/flash-arts/": {
      "get": {
        "tags": [
          "flash-arts"
        ],
        "summary": "Listar Flash Arts",
        "operationId": "listar_flash_arts_api_v1_flash_arts__get",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlashArtResponse"
                  },
                  "title": "Response Listar Flash Arts Api V1 Flash Arts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "flash-arts"
        ],
        "summary": "Criar Flash Art",
        "operationId": "criar_flash_art_api_v1_flash_arts__post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_criar_flash_art_api_v1_flash_arts__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashArtResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flash-arts/{id}/imagem": {
      "get": {
        "tags": [
          "flash-arts"
        ],
        "summary": "Servir Imagem",
        "description": "Serve imagem autenticada da flash art.",
        "operationId": "servir_imagem_api_v1_flash_arts__id__imagem_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flash-arts/{id}/status": {
      "patch": {
        "tags": [
          "flash-arts"
        ],
        "summary": "Alterar Status Flash",
        "operationId": "alterar_status_flash_api_v1_flash_arts__id__status_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "status_novo",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StatusFlash"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashArtResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flash-arts/{id}": {
      "delete": {
        "tags": [
          "flash-arts"
        ],
        "summary": "Arquivar Flash Art",
        "operationId": "arquivar_flash_art_api_v1_flash_arts__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/{slug}/flash-arts": {
      "get": {
        "tags": [
          "portal-público"
        ],
        "summary": "Flash Arts Publicas",
        "description": "Retorna flash arts disponíveis do estúdio — sem autenticação.",
        "operationId": "flash_arts_publicas_api_v1_public__slug__flash_arts_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlashArtPublicaResponse"
                  },
                  "title": "Response Flash Arts Publicas Api V1 Public  Slug  Flash Arts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/{slug}/flash-arts/{flash_id}/imagem": {
      "get": {
        "tags": [
          "portal-público"
        ],
        "summary": "Imagem Flash Art Publica",
        "description": "Serve imagem pública de flash art — sem autenticação.",
        "operationId": "imagem_flash_art_publica_api_v1_public__slug__flash_arts__flash_id__imagem_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "flash_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Flash Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_criar_flash_art_api_v1_flash_arts__post": {
        "properties": {
          "arquivo": {
            "type": "string",
            "format": "binary",
            "title": "Arquivo"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco"
          },
          "tamanho_sugerido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tamanho Sugerido"
          },
          "local_recomendado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Recomendado"
          }
        },
        "type": "object",
        "required": [
          "arquivo",
          "titulo"
        ],
        "title": "Body_criar_flash_art_api_v1_flash_arts__post"
      },
...
      "FlashArtPublicaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco"
          },
          "tamanho_sugerido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tamanho Sugerido"
          },
          "local_recomendado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Recomendado"
          },
          "has_imagem": {
            "type": "boolean",
            "title": "Has Imagem"
          }
        },
        "type": "object",
        "required": [
          "id",
          "titulo",
          "descricao",
          "preco",
          "tamanho_sugerido",
          "local_recomendado",
          "has_imagem"
        ],
        "title": "FlashArtPublicaResponse"
      },
      "FlashArtResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "estudio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Estudio Id"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "imagem_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagem Path"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco"
          },
          "tamanho_sugerido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tamanho Sugerido"
          },
          "local_recomendado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Recomendado"
          },
          "status": {
            "$ref": "#/components/schemas/StatusFlash"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "estudio_id",
          "titulo",
          "descricao",
          "imagem_path",
          "preco",
          "tamanho_sugerido",
          "local_recomendado",
          "status",
          "ativo"
        ],
        "title": "FlashArtResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusFlash": {
        "type": "string",
        "enum": [
          "DISPONIVEL",
          "EM_NEGOCIACAO",
          "RESERVADA",
          "SINAL_PAGO",
          "VENDIDA",
          "ARQUIVADA"
        ],
        "title": "StatusFlash"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}